[IMP] l10n_es_facturae: Make it work for physical people
