+++ l10n_es_facturae/i18n/es.po
@@ -715,24 +715,6 @@ msgstr "Otro"
 #. module: l10n_es_facturae
 #: code:addons/l10n_es_facturae/models/account_move.py:0
 #, python-format
-msgid "Oficina Contable not provided"
-msgstr "La Oficina Contable es obligatorio"
-
-#. module: l10n_es_facturae
-#: code:addons/l10n_es_facturae/models/account_move.py:0
-#, python-format
-msgid "Organo Gestor not provided"
-msgstr "El Organo Gestor es obligatorio"
-
-#. module: l10n_es_facturae
-#: code:addons/l10n_es_facturae/models/account_move.py:0
-#, python-format
-msgid "Unidad Tramitadora not provided"
-msgstr "La Unidad Tramitadora es obligatorio"
-
-#. module: l10n_es_facturae
-#: code:addons/l10n_es_facturae/models/account_move.py:0
-#, python-format
 msgid "Partner state not provided"
 msgstr "El estado es obligatorio"
