+++ l10n_es_facturae/models/account_move.py
@@ -148,12 +148,6 @@ class AccountMove(models.Model):
             raise ValidationError(_("Partner vat is too small"))
         if not self.partner_id.state_id:
             raise ValidationError(_("Partner state not provided"))
-        if not self.partner_id.organo_gestor:
-            raise ValidationError(_("Organo Gestor not provided"))
-        if not self.partner_id.unidad_tramitadora:
-            raise ValidationError(_("Unidad Tramitadora not provided"))
-        if not self.partner_id.oficina_contable:
-            raise ValidationError(_("Oficina Contable not provided"))
         if not self.payment_mode_id:
             raise ValidationError(_("Payment mode is required"))
         if self.payment_mode_id.facturae_code:
