+++ l10n_es_facturae_face/i18n/l10n_es_facturae_face.pot
@@ -228,6 +228,18 @@ msgid "OK"
 msgstr ""
 
 #. module: l10n_es_facturae_face
+#: code:addons/l10n_es_facturae_face/models/account_move.py:0
+#, python-format
+msgid "Oficina Contable not provided"
+msgstr ""
+
+#. module: l10n_es_facturae_face
+#: code:addons/l10n_es_facturae_face/models/account_move.py:0
+#, python-format
+msgid "Organo Gestor not provided"
+msgstr ""
+
+#. module: l10n_es_facturae_face
 #: model:ir.model.fields.selection,name:l10n_es_facturae_face.selection__account_move__l10n_es_facturae_status__face-2500
 msgid "Paid"
 msgstr ""
@@ -257,6 +269,12 @@ msgstr ""
 #: code:addons/l10n_es_facturae_face/models/res_partner.py:0
 #, python-format
 msgid "State must be defined in Spain in order to send to FACe"
+msgstr ""
+
+#. module: l10n_es_facturae_face
+#: code:addons/l10n_es_facturae_face/models/account_move.py:0
+#, python-format
+msgid "Unidad Tramitadora not provided"
 msgstr ""
 
 #. module: l10n_es_facturae_face
