+++ l10n_es_facturae_face/models/account_move.py
@@ -2,7 +2,8 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
 
-from odoo import api, fields, models
+from odoo import _, api, fields, models
+from odoo.exceptions import ValidationError
 
 
 class AccountMove(models.Model):
@@ -83,3 +84,16 @@ class AccountMove(models.Model):
             "face-2600",
             "face-3100",
         ]
+
+    def validate_facturae_fields(self):
+        super().validate_facturae_fields()
+        if (
+            self.partner_id.l10n_es_facturae_sending_code == "face"
+            and not self.partner_id.organo_gestor
+        ):
+            raise ValidationError(_("Organo Gestor not provided"))
+        if not self.partner_id.unidad_tramitadora:
+            raise ValidationError(_("Unidad Tramitadora not provided"))
+        if not self.partner_id.oficina_contable:
+            raise ValidationError(_("Oficina Contable not provided"))
+        return
